third_party/blink/web_tests/http/tests/security/mixedContent/empty-url-plugin-in-frame.html

<html>
<body>
<script>
if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.waitUntilDone();
    window.addEventListener("message", _ => {
      testRunner.notifyDone();
    });
}
</script>
<p>This test loads a secure iframe that loads a plugin without a URL.  We should
*not* get a mixed content callback because the plugin cannot be controlled by
an active network attacker.</p>
<iframe src="https://127.0.0.1:8443/security/mixedContent/resources/frame-with-empty-url-plugin.html"></iframe>
</body>
</html>
