third_party/blink/web_tests/http/tests/security/mixedContent/empty-url-plugin-in-frame-expected.txt

CONSOLE MESSAGE: Blink Test Plugin: initializing
This test loads a secure iframe that loads a plugin without a URL. We should *not* get a mixed content callback because the plugin cannot be controlled by an active network attacker.
